home page to separate class
